+++ cookbooks/graphics/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of graphics.
 
+## 0.2.0:
+
+* recipes/default - now includes pencil install management
+* recipes/pencil  - manages pencil app installation
+
 ## 0.1.0:
 
 * Initial release of graphics